Update and rename eb.js to electric_bus.js
Revised based on suggestions from mark m.
